+++ kotlin-stdlib/commonMain/kotlin/com/javiersc/kotlin/stdlib/Strings.kt
@@ -37,4 +37,5 @@ public inline fun String?.isNotNullNorEmpty(): Boolean {
     return (this != null) && isNotEmpty()
 }
 
-public inline fun emptyString(): String = ""
+public inline val String.Companion.Empty: String
+    get() = ""